change namespace and class name
