+++ lib/middleman-robots/generators/block.rb
@@ -1,0 +1,43 @@
+module Middleman
+  module Robots
+    module Generators
+      # Block
+      #
+      # Generating Block in robots.txt
+      class Block
+        attr_accessor :rule
+
+        def initialize(rule)
+          @rule = rule
+        end
+
+        def text
+          [
+            user_agent,
+            disallow,
+            allow,
+          ].compact.join("\n")
+        end
+
+        private
+
+        def user_agent
+          user_agent = rule[:user_agent] || rule['user-agent']  || '*'
+          "User-Agent: #{user_agent}"
+        end
+
+        def disallow
+          return nil unless rule.key? :disallow
+
+          rule[:disallow].map { |path| "Disallow: #{File.join('/', path)}" }
+        end
+
+        def allow
+          return nil unless rule.key? :allow
+
+          rule[:allow].map { |path| "Allow: #{File.join('/', path)}" }
+        end
+      end
+    end
+  end
+end
