+++ lib/middleman-robots/generators/sitemap_uri.rb
@@ -1,0 +1,19 @@
+module Middleman
+  module Robots
+    module Generators
+      class SitemapUri
+        attr_accessor :uri
+
+        def initialize(uri)
+          @uri = uri
+        end
+
+        def text
+          return nil if uri.nil?
+
+          "Sitemap: #{uri}"
+        end
+      end
+    end
+  end
+end
