--- lib/middleman-robots/sitemap_uri.rb
@@ -1,17 +1,0 @@
-module Middleman
-  module Robots
-    class SitemapUri
-      attr_accessor :uri
-
-      def initialize(uri)
-        @uri = uri
-      end
-
-      def text
-        return nil if uri.nil?
-
-        "Sitemap: #{uri}"
-      end
-    end
-  end
-end
